@@ -8,54 +8,68 @@
     "test": "ng test",
     "coverage": "ng test --code-coverage --sourceMap=true",
     "lint": "ng lint --type-check --format stylish",
+    "lint:fix": "ng lint --type-check --format stylish --fix",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/http": "~7.1.0",
-    "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "bootstrap": "^4.3.1",
-    "core-js": "2.5.4",
+    "@angular/animations": "^13.3.1",
+    "@angular/cdk": "^13.3.2",
+    "@angular/common": "^13.3.1",
+    "@angular/compiler": "^13.3.1",
+    "@angular/core": "^13.3.1",
+    "@angular/forms": "^13.3.1",
+    "@angular/material-moment-adapter": "^13.3.2",
+    "@angular/platform-browser": "^13.3.1",
+    "@angular/platform-browser-dynamic": "^13.3.1",
+    "@angular/router": "^13.3.1",
+    "@js-temporal/polyfill": "^0.4.0",
+    "@ngtools/webpack": "^13.3.1",
+    "@popperjs/core": "^2.11.5",
+    "ansi-regex": "^6.0.1",
+    "bootstrap": "^5.1.3",
+    "core-js": "^2.6.12",
+    "eslint": "^8.12.0",
     "font-awesome": "^4.7.0",
-    "rxjs": "~6.3.3",
-    "socket.io": "^2.2.0",
+    "ini": "^3.0.0",
+    "jasmine": "^4.0.2",
+    "jasminewd2": "^2.2.0",
+    "jquery": "^3.6.0",
+    "moment": "^2.29.1",
+    "reflect": "^0.1.3",
+    "rxjs": "^7.5.5",
+    "socket.io": "^4.4.1",
     "socket.io-client": "^2.2.0",
-    "three": "^0.93.0",
-    "zone.js": "~0.8.26"
+    "three": "^0.139.0",
+    "webpack": "^5.71.0",
+    "xmlhttprequest-ssl": "^2.0.0",
+    "yargs-parser": "^21.0.1",
+    "zone.js": "^0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.4",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "~7.1.0",
+    "@angular-devkit/build-angular": "^13.3.1",
+    "@angular/cli": "^13.3.1",
+    "@angular/compiler-cli": "^13.3.1",
+    "@angular/language-service": "~8.2.14",
+    "@angular/material": "^13.3.2",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.12.18",
-    "@types/socket.io": "^2.1.2",
-    "@types/socket.io-client": "^1.4.32",
-    "@types/three": "^0.93.16",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.11.0",
-    "tslint-microsoft-contrib": "6.0.0",
-    "typescript": "~3.1.0"
+    "@types/node": "^16.11.26",
+    "@types/three": "^0.139.0",
+    "coercion": "^1.1.0",
+    "hammerjs": "^2.0.8",
+    "jasmine-core": "^4.0.1",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.3.17",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-firefox-launcher": "^2.1.2",
+    "karma-jasmine": "^4.0.2",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "platform": "^1.3.6",
+    "protractor": "^7.0.0",
+    "sass": "^1.50.0",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
   }
 }
